DEMOS-2286: comp'd create new and moved to title
In DefaultHeaderLower there's the createNew buttons. They inc the create
demo ext and amendment buttons, but this is on the wrong line in the
header. Was able to reduce the header lower page and moved to Header
file
